+++ semana18/package.json
@@ -17,7 +17,7 @@
     "@types/uuid": "^8.3.0",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
-    "express": "^4.17.1",
+    "express": "^4.17.0",
     "jsonwebtoken": "^8.5.1",
     "knex": "^0.95.6",
     "mysql": "^2.18.1",
@@ -25,7 +25,7 @@
   },
   "devDependencies": {
     "@types/cors": "^2.8.10",
-    "@types/express": "^4.17.12",
+    "@types/express": "^4.17.0",
     "ts-node-dev": "^1.1.6",
     "typescript": "^4.3.2"
   }
