+++ semana18/services/idGenerator.ts
@@ -1,0 +1,5 @@
+import {v4} from "uuid";
+
+export function generatedId(): string{
+    return v4();
+}
